Backend/requirements.txt
aenum
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
boto3
certifi
charset-normalizer
click
colorama
dataclasses-json
deepgram-sdk
deprecation
distro
fastapi
frozenlist
greenlet
groq
h11
httpcore
httpx
idna
jiter
jsonpatch
jsonpointer
langchain
langchain-core
langchain-groq
langchain-openai
langchain-text-splitters
langsmith
marshmallow
multidict
mypy_extensions
nltk
numpy
openai
orjson
packaging
pandas
pillow
propcache
pydantic
pydantic_core
pymupdf
python-docx
python-dotenv
python-multipart
pytesseract
PyYAML
regex
requests
requests-toolbelt
sentence-transformers
sniffio
snowflake-connector-python
SQLAlchemy
starlette
tenacity
tiktoken
tqdm
typing-inspect
typing-inspection
typing_extensions
urllib3
uvicorn
websockets
yarl
zstandard
